python-requirement.txt
Django==1.7.4
MySQL-python==1.2.5
Pillow==2.7.0
django-localflavor==1.1
djangorestframework==3.0.4
psycopg2==2.5.4
uWSGI==2.0.9
wsgiref==0.1.2
